src/main/java/bridge/view/InputView.java
package bridge.view;

import camp.nextstep.edu.missionutils.Console;

/**
 * 사용자로부터 입력을 받는 역할을 한다.
 */
public class InputView {

    /**
     * 다리의 길이를 입력받는다.
     */
    public int readBridgeSize(){
        String bridgeSizeInput = Console.readLine().trim();
        InputValidator.validateBridgeSize(bridgeSizeInput);
        return Integer.parseInt(bridgeSizeInput);
    }

    /**
     * 사용자가 이동할 칸을 입력받는다.
     */
    public String readMoving() {
        String movingDirection = Console.readLine().trim();
        InputValidator.validateMovingDirection(movingDirection);
        return movingDirection;
    }

    /**
     * 사용자가 게임을 다시 시도할지 종료할지 여부를 입력받는다.
     */
    public String readGameCommand() {
        String gameCommand = Console.readLine().trim();
        InputValidator.validateGameCommand(gameCommand);
        return gameCommand;
    }
}
